@@ -1,14 +1,10 @@
 package com.eserviceplatform.mobile.camerasupport.plugin.view;
 
-import android.content.Context;
-import android.graphics.SurfaceTexture;
-import android.support.v4.view.ViewCompat;
-import android.util.Size;
+import androidx.core.view.ViewCompat;
+
 import android.view.Surface;
 import android.view.SurfaceHolder;
-import android.view.SurfaceView;
 import android.view.View;
-import android.view.ViewGroup;
 
 import com.eserviceplatform.mobile.camerasupport.plugin.base.PreviewImpl;
 
